Improve stories
